+++ pysui/sui/sui_pgql/pgql_txn_argb.py
@@ -37,6 +37,15 @@ _SCALARS_BCS = {
     "u256": bcs.U256,
 }
 
+_OPTIONAL_SCALARS_BCS = {
+    "u8": bcs.OptionalU8,
+    "u16": bcs.OptionalU16,
+    "u32": bcs.OptionalU32,
+    "u64": bcs.OptionalU64,
+    "u128": bcs.OptionalU128,
+    "u256": bcs.OptionalU256,
+}
+
 
 @dataclass
 class _ArgSummary:
@@ -125,11 +134,13 @@ class ResolvingArgParser:
                             ),
                         ),
                     )
-
-                return (
-                    _SCALARS.get(expected_type.scalar_type),
-                    PureInput.as_input,
-                )
+                if in_optional:
+                    return _OPTIONAL_SCALARS_BCS.get(expected_type.scalar_type)
+                else:
+                    return (
+                        _SCALARS.get(expected_type.scalar_type),
+                        PureInput.as_input,
+                    )
 
     def _scalar_argument(
         self,
@@ -345,6 +356,8 @@ class ResolvingArgParser:
                 etr = bcs.OptionalTypeFactory.as_optional(inner_type)
         elif construct:
             inner_type = construct(arg)
+            if construct in bcs.OPTIONAL_SCALARS:
+                return inner_type
             etr = bcs.OptionalTypeFactory.as_optional(inner_type)
         etr.value = inner_type
         return etr