+++ pysui/sui/sui_txn/transaction_builder.py
@@ -453,8 +453,10 @@ class ProgrammableTransactionBuilder:
                 argrefs.append(self.input_pure(PureInput.as_input(arg)))
             elif isinstance(arg, tuple):
                 argrefs.append(self.input_obj(*arg))
-            elif isinstance(arg, (bcs.Argument, bcs.OptionalU64)):
+            elif isinstance(arg, bcs.Argument):
                 argrefs.append(arg)
+            elif isinstance(arg, tuple(bcs.OPTIONAL_SCALARS)):
+                argrefs.append(self.input_pure(PureInput.as_input(arg)))
             elif isinstance(arg, list):
                 argrefs.append(self.input_pure(PureInput.as_input(arg)))
             elif isinstance(arg, bcs.Variable):
@@ -574,7 +576,7 @@ class ProgrammableTransactionBuilder:
         self,
         recipient: bcs.BuilderArg,
         from_coin: Union[bcs.Argument, tuple[bcs.BuilderArg, bcs.ObjectArg]],
-        amount: Optional[Union[bcs.BuilderArg, bcs.Optional]] = None,
+        amount: Optional[Union[bcs.BuilderArg, bcs.OptionalU64]] = None,
     ) -> bcs.Argument:
         """Setup a TransferObjects for Sui Coins.
 
@@ -584,12 +586,9 @@ class ProgrammableTransactionBuilder:
         reciever_arg = self.input_pure(recipient)
         if amount and isinstance(amount, bcs.BuilderArg):
             coin_arg = self.split_coin(from_coin=from_coin, amounts=[amount])
-        elif isinstance(amount, bcs.Optional):
-            o_amount: bcs.Optional = amount
-            if o_amount.value:
-                coin_arg = self.split_coin(
-                    from_coin=from_coin, amounts=[o_amount.value]
-                )
+        elif isinstance(amount, bcs.OptionalU64):
+            if amount.value:
+                coin_arg = self.split_coin(from_coin=from_coin, amounts=[amount.value])
             else:
                 coin_arg = from_coin
         else: