+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.75.0] - Unpublished
+## [0.75.0] - 2025-01-03
 
 **BREAKING CHANGES to SerialTransactionExecution constructor**
 
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [change](https://github.com/FrankC01/pysui/issues/253) SerialTransactionExecution keywords on init changes
 - [change](https://github.com/FrankC01/pysui/issues/252) SerialTransactionExectuor use int balance in _smash_gas sort
+- Bumped pysui-fastcrypto to 0.5.1
 
 ### Removed